refactoring usecase name
