+++ src/main/java/use_case/blackjack/hit/BlackjackGetCardInputData.java
@@ -1,4 +1,4 @@
-package use_case.blackjack.get_card;
+package use_case.blackjack.hit;
 
 /**
  * Input Data for the Blackjack Get Card Use Case.
